move internal gateway controller to local package
