+++ src/main/java/org/semantics/apigateway/controller/local/CollectionsController.java
@@ -1,9 +1,6 @@
-package org.semantics.apigateway.controller;
+package org.semantics.apigateway.controller.local;
 
-import io.swagger.v3.oas.annotations.enums.SecuritySchemeIn;
-import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
-import io.swagger.v3.oas.annotations.security.SecurityScheme;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
 import org.semantics.apigateway.model.user.TerminologyCollection;
