+++ src/main/java/org/semantics/apigateway/controller/local/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package org.semantics.apigateway.controller;
+package org.semantics.apigateway.controller.local;
 
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.MalformedJwtException;
